♻️ move bus implementation in dedicated module
