+++ marketplace-indexer/src/infrastructure/event_bus/observer.rs
@@ -2,14 +2,14 @@ use crate::domain::{BlockchainObserver, ObservedEvent};
 use async_trait::async_trait;
 use log::error;
 use marketplace_domain::{Destination, EventOrigin, Publisher};
-use marketplace_event_store::{event_bus, Event};
+use marketplace_event_store::{bus::QUEUE_NAME as EVENT_STORE_QUEUE, Event};
 
 #[async_trait]
 impl<P: Publisher<Event>> BlockchainObserver for P {
 	async fn on_new_event(&self, observed_event: &ObservedEvent, _block_number: u64) {
 		if let Err(error) = self
 			.publish(
-				Destination::queue(event_bus::QUEUE_NAME),
+				Destination::queue(EVENT_STORE_QUEUE),
 				&observed_event.clone().into(),
 			)
 			.await
