+++ marketplace-infrastructure/src/amqp/event_bus.rs
@@ -1,10 +1,10 @@
-use super::{Bus, ConsumableBus, Error};
+use super::{Bus, BusError, ConsumableBus};
 use lapin::options::QueueDeclareOptions;
 use log::info;
 
 pub const EXCHANGE_NAME: &str = "events";
 
-pub async fn consumer() -> Result<ConsumableBus, Error> {
+pub async fn consumer() -> Result<ConsumableBus, BusError> {
 	let event_bus = Bus::default()
 		.await?
 		.with_queue(
